backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]==3.3.0
passlib[bcrypt]
python-dotenv
pyyaml
pydantic==2.7.1
pydantic[email]
playwright==1.44.0
anthropic>=0.39.0
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
cryptography==42.0.5
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dateutil==2.9.0
requests
bcrypt==4.1.3


